move detailed balance to utils
