+++ src/easydynamics/utils/__init__.py
@@ -1,0 +1,3 @@
+from .utils import detailed_balance_factor
+
+__all__ = ['detailed_balance_factor']
